fix: all commands go through the queue
There were some instances where any command that responded with a `+OK`
and not a value bypassed the command queue. This caused
issues, as the subsequent `GET` command to update feedback for a router
output WAS enqueued, and responses from subsequent `SET` responses were
stepping on the `+OK 0` responses to the `GET` commands. This is
especially apparent on RS-232 connections, and probably doesn't present
as an issue for SSH connections.

ref: #89