fix: serialization of TesiraPreset works correctly
The `TesiraPreset` class inherits from the `TesiraDspPresets` class,
which has a `StringFeedback` for use with bridging to SIMPL. The
feedback was causing serialization issues when attempting to send
presets to a React application via MC.

The `TesiraPreset` class also had an extra `TesiraDspPresets` type that
wasn't necassary, since the class inherits from `TesiraDspPresets` and
has all the same properties due to the inheritance.

The `SERIES4` directive was removed, and the log statements in the `TesiraDsp` class and the `TesiraDspPreset` class were changed to use the extension methods.