fix: make volume repeat rate configurable
In some situations, the volume up/down commands for a fader were
getting queued every 100ms, while they were getting sent to the DSP
every 250ms. The repeat time for the volume up/down commands is now configurable per fader
with a default of 250ms to match up with the rate for the transmit queue.